requirements.txt
uv
google-adk
python-dotenv
praw